# board-service/docs/swagger.yaml
basePath: /api
definitions:
...
  project-board-api_internal_dto.CreateProjectRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      workspaceId:
        type: string
    required:
    - name
    - workspaceId
    type: object
  project-board-api_internal_dto.FieldOption:
    properties:
      optionId:
        type: string
      optionLabel:
        type: string
      optionValue:
        type: string
    type: object
...
  project-board-api_internal_dto.FieldTypeInfo:
    properties:
      description:
        type: string
      typeId:
        type: string
      typeName:
        type: string
    type: object
  project-board-api_internal_dto.FieldWithOptionsResponse:
    properties:
      description:
        type: string
      fieldId:
        type: string
      fieldName:
        type: string
      fieldType:
        type: string
      isRequired:
        type: boolean
      options:
        items:
          $ref: '#/definitions/project-board-api_internal_dto.FieldOption'
        type: array
    type: object
...
  project-board-api_internal_dto.PaginatedProjectsResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      projects:
        items:
          $ref: '#/definitions/project-board-api_internal_dto.ProjectResponse'
        type: array
      total:
        type: integer
    type: object
...
  project-board-api_internal_dto.ProjectBasicInfo:
    properties:
      createdAt:
        type: string
      description:
        type: string
      isPublic:
        type: boolean
      name:
        type: string
      ownerId:
        type: string
      projectId:
        type: string
      updatedAt:
        type: string
      workspaceEmail:
        type: string
      workspaceId:
        type: string
      workspaceName:
        type: string
    type: object
  project-board-api_internal_dto.ProjectInitSettingsResponse:
    properties:
      defaultViewId:
        type: string
      fieldTypes:
        items:
          $ref: '#/definitions/project-board-api_internal_dto.FieldTypeInfo'
        type: array
      fields:
        items:
          $ref: '#/definitions/project-board-api_internal_dto.FieldWithOptionsResponse'
        type: array
      project:
        $ref: '#/definitions/project-board-api_internal_dto.ProjectBasicInfo'
    type: object
...
  project-board-api_internal_dto.ProjectResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      isPublic:
        type: boolean
      name:
        type: string
      ownerEmail:
        type: string
      ownerId:
        type: string
      ownerName:
        type: string
      projectId:
        type: string
      updatedAt:
        type: string
      workspaceId:
        type: string
    type: object
...
  project-board-api_internal_dto.UpdateProjectRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    type: object
  project-board-api_internal_response.ErrorResponse:
    properties:
      error: {}
      requestId:
        type: string
    type: object
  project-board-api_internal_response.SuccessResponse:
    properties:
      data: {}
      requestId:
        type: string
    type: object
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    프로젝트 보드 관리 시스템 API 서버입니다.
    Board, Project, Comment, Participant 관리 기능을 제공합니다.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Project Board Management API
  version: "1.0"
paths:
...
  /projects:
    get:
      description: 특정 Workspace에 속한 모든 Project를 조회합니다. 프론트엔드 호환용 엔드포인트
      parameters:
      - description: Workspace ID (UUID)
        in: query
        name: workspaceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project 목록 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.PaginatedProjectsResponse'
              type: object
        "400":
          description: 잘못된 Workspace ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Workspace의 Project 목록 조회 (쿼리 파라미터 방식)
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: 새로운 Project를 생성합니다
      parameters:
      - description: Project 생성 요청
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project-board-api_internal_dto.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Project 생성 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.ProjectResponse'
              type: object
        "400":
          description: 잘못된 요청
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Project 생성
      tags:
      - projects
  /projects/{projectId}:
    delete:
      description: Project를 삭제합니다 (OWNER만 가능)
      parameters:
      - description: Project ID (UUID)
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project 삭제 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: 잘못된 Project ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "403":
          description: 권한 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Project를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Project 삭제
      tags:
      - projects
    get:
      description: 특정 Project의 상세 정보를 조회합니다
      parameters:
      - description: Project ID (UUID)
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.ProjectResponse'
              type: object
        "400":
          description: 잘못된 Project ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "403":
          description: 권한 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Project를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Project 상세 조회
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Project의 이름과 설명을 수정합니다 (OWNER만 가능)
      parameters:
      - description: Project ID (UUID)
        in: path
        name: projectId
        required: true
        type: string
      - description: Project 수정 요청
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project-board-api_internal_dto.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Project 수정 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.ProjectResponse'
              type: object
        "400":
          description: 잘못된 요청
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "403":
          description: 권한 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Project를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Project 수정
      tags:
      - projects
  /projects/{projectId}/init-settings:
    get:
      description: Project 진입 시 필요한 초기 설정 데이터를 한 번에 조회합니다
      parameters:
      - description: Project ID (UUID)
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 초기 설정 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.ProjectInitSettingsResponse'
              type: object
        "400":
          description: 잘못된 Project ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "403":
          description: 권한 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Project를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Project 초기 설정 조회
      tags:
      - projects
...
  /projects/search:
    get:
      description: Workspace 내에서 Project를 이름이나 설명으로 검색합니다
      parameters:
      - description: Workspace ID (UUID)
        in: query
        name: workspaceId
        required: true
        type: string
      - description: 검색어
        in: query
        name: query
        required: true
        type: string
      - default: 1
        description: 페이지 번호
        in: query
        name: page
        type: integer
      - default: 10
        description: 페이지 크기
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 검색 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.PaginatedProjectsResponse'
              type: object
        "400":
          description: 잘못된 요청
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "403":
          description: 권한 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Project 검색
      tags:
      - projects
  /projects/workspace/{workspaceId}:
    get:
      description: 특정 Workspace에 속한 모든 Project를 조회합니다
      parameters:
      - description: Workspace ID (UUID)
        in: path
        name: workspaceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project 목록 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/project-board-api_internal_dto.ProjectResponse'
                  type: array
              type: object
        "400":
          description: 잘못된 Workspace ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Workspace의 Project 목록 조회
      tags:
      - projects
  /projects/workspace/{workspaceId}/default:
    get:
      description: 특정 Workspace의 기본(default) Project를 조회합니다
      parameters:
      - description: Workspace ID (UUID)
        in: path
        name: workspaceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 기본 Project 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.ProjectResponse'
              type: object
        "400":
          description: 잘못된 Workspace ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: 기본 Project를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Workspace의 기본 Project 조회
      tags:
      - projects
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT 토큰을 입력하세요. 형식: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"